# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: E-Commerce-API
  contact: {}
  version: "1.0"
servers:
  - url: https://e-commerce-kkkmaks-g7aal.ondigitalocean.app/api/v1
    variables: {}
paths:
...
  /auth/verify-email:
    post:
      tags:
        - Auth
      summary: Verify Email
      operationId: VerifyEmail
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/VerifyEmailRequest"
                - example:
                    verificationToken: ccc945da001a58c8eaf59066e8da0888dfd6ada165c372a45cea047c2a5274c00edf3ec1ac8177c890900c97
                    email: kurchenko.maks.fb04@gmail.com
            example:
              verificationToken: ccc945da001a58c8eaf59066e8da0888dfd6ada165c372a45cea047c2a5274c00edf3ec1ac8177c890900c97
              email: kurchenko.maks.fb04@gmail.com
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
  /auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Forgot Password
      operationId: ForgotPassword
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ForgotPasswordRequest"
                - example:
                    email: kurchenko.maks.fb04@gmail.com
            example:
              email: kurchenko.maks.fb04@gmail.com
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset Password
      operationId: ResetPassword
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ResetPasswordRequest"
                - example:
                    resetPasswordToken: 282fe0c10d3ee6e4243c17016ccbb7c05419d62fa2a94a5a1767df74a3766033dcd3054c7dd54e540e1ed08d
                    email: kurchenko.maks.fb04@gmail.com
                    password1: testreset
                    password2: testreset
            example:
              resetPasswordToken: 282fe0c10d3ee6e4243c17016ccbb7c05419d62fa2a94a5a1767df74a3766033dcd3054c7dd54e540e1ed08d
              email: kurchenko.maks.fb04@gmail.com
              password1: testreset
              password2: testreset
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    VerifyEmailRequest:
      title: VerifyEmailRequest
      required:
        - verificationToken
        - email
      type: object
      properties:
        verificationToken:
          type: string
        email:
          type: string
      example:
        verificationToken: ccc945da001a58c8eaf59066e8da0888dfd6ada165c372a45cea047c2a5274c00edf3ec1ac8177c890900c97
        email: kurchenko.maks.fb04@gmail.com
    ForgotPasswordRequest:
      title: ForgotPasswordRequest
      required:
        - email
      type: object
      properties:
        email:
          type: string
      example:
        email: kurchenko.maks.fb04@gmail.com
    ResetPasswordRequest:
      title: ResetPasswordRequest
      required:
        - resetPasswordToken
        - email
        - password1
        - password2
      type: object
      properties:
        resetPasswordToken:
          type: string
        email:
          type: string
        password1:
          type: string
        password2:
          type: string
      example:
        resetPasswordToken: 282fe0c10d3ee6e4243c17016ccbb7c05419d62fa2a94a5a1767df74a3766033dcd3054c7dd54e540e1ed08d
        email: kurchenko.maks.fb04@gmail.com
        password1: testreset
        password2: testreset
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
...